+++ src/server/devices/deviceUpdater.ts
@@ -37,22 +37,25 @@ export class DeviceUpdater implements IDeviceUpdater {
       classifier = Devices.IDENTIFIER_ZIGBEE;
       devId = idSplit[2].substring(2);
     } else if (idSplit[0] == Devices.IDENTIFIER_SMART_GARDEN) {
-      ServerLogService.writeLog(LogLevel.Debug, `Received smart-garden state (${id}): ${JSON.stringify(state)}`);
       if (
         idSplit[1] === 'admin' ||
         idSplit.length < 4 ||
         idSplit[2].indexOf('LOCATION') !== 0 ||
         idSplit[3].indexOf('DEVICE') !== 0
       ) {
-        ServerLogService.writeLog(LogLevel.Debug, `Discarded smart-garden state id-split: ${JSON.stringify(idSplit)}`);
         // This is no update for a smartgarden device
         return;
       }
       classifier = Devices.IDENTIFIER_SMART_GARDEN;
       devId = idSplit[3];
     }
-    const device: undefined | iBaseDevice = API.getDevice(`${classifier}-${devId}`, false);
+    const allDevicesKey: string = `${classifier}-${devId}`;
+    const device: undefined | iBaseDevice = API.getDevice(allDevicesKey, false);
     if (typeof device === 'undefined' || (device as IoBrokerBaseDevice).update === undefined) {
+      ServerLogService.writeLog(
+        LogLevel.Error,
+        `deviceUpdater.updateState('${id}', ${JSON.stringify(state)}'): Device ${allDevicesKey} not found`,
+      );
       return;
     }
     try {